% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conv_hr.R
\name{C_to_F}
\alias{C_to_F}
\title{Celsius to Fahrenheit Conversion}
\usage{
C_to_F(C_temp)
}
\arguments{
\item{C_temp}{Numeric vector representing temperature(s) in degrees Celsius.}
}
\value{
Numeric vector representing temperature(s) in degrees Fahrenheit.
}
\description{
Convert temperature from degrees Celsius to degrees Fahrenheit.
}
\examples{
C_to_F(20)
C_to_F(c(-10, 0, 25))
}
