% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conv_hr.R
\name{C_to_K}
\alias{C_to_K}
\title{Celsius to Kelvin Conversion}
\usage{
C_to_K(C_temp)
}
\arguments{
\item{C_temp}{Numeric vector representing temperature(s) in degrees Celsius.}
}
\value{
Numeric vector representing temperature(s) in Kelvin.
}
\description{
Convert temperature from degrees Celsius to Kelvin.
}
\examples{
C_to_K(0)
C_to_K(c(-10, 25, 100))
}
