% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conv_hr.R
\name{F_to_C}
\alias{F_to_C}
\title{Fahrenheit to Celsius Conversion}
\usage{
F_to_C(F_temp)
}
\arguments{
\item{F_temp}{Numeric vector representing temperature(s) in degrees Fahrenheit.}
}
\value{
Numeric vector representing temperature(s) in degrees Celsius.
}
\description{
Convert temperature from degrees Fahrenheit to degrees Celsius.
}
\examples{
F_to_C(50)
F_to_C(c(32, 68, 104))
}
