% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conv_hr.R
\name{F_to_K}
\alias{F_to_K}
\title{Fahrenheit to Kelvin Conversion}
\usage{
F_to_K(F_temp)
}
\arguments{
\item{F_temp}{Numeric vector representing temperature(s) in degrees Fahrenheit.}
}
\value{
Numeric vector representing temperature(s) in Kelvin.
}
\description{
Convert temperature from degrees Fahrenheit to Kelvin.
}
\examples{
F_to_K(32)
F_to_K(c(68, 100))
}
