% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conv_hr.R
\name{K_to_C}
\alias{K_to_C}
\title{Kelvin to Celsius Conversion}
\usage{
K_to_C(K_temp)
}
\arguments{
\item{K_temp}{Numeric vector representing temperature(s) in Kelvin.}
}
\value{
Numeric vector representing temperature(s) in degrees Celsius.
}
\description{
Convert temperature from Kelvin to degrees Celsius.
}
\examples{
K_to_C(273)
K_to_C(c(300, 250))
}
