% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conv_hr.R
\name{K_to_F}
\alias{K_to_F}
\title{Kelvin to Fahrenheit Conversion}
\usage{
K_to_F(K_temp)
}
\arguments{
\item{K_temp}{Numeric vector representing temperature(s) in Kelvin.}
}
\value{
Numeric vector representing temperature(s) in degrees Fahrenheit.
}
\description{
Convert temperature from Kelvin to degrees Fahrenheit.
}
\examples{
K_to_F(273)
K_to_F(c(300, 310))
}
